+++ herencia/CuentaAhorro.js
@@ -1,2 +1,4 @@
-export class CuentaAhorro {
-  }
+import { Cuenta } from "./Cuenta.js"
+export class CuentaAhorro extends Cuenta {
+
+}
