+++ herencia/index.js
@@ -1,16 +1,15 @@
 /* import class */
 import { Cliente } from "./Cliente.js";
-// import { CuentaCorriente } from "./CuentaCorriente.js";
-// import { CuentaAhorro } from './CuentaAhorro.js';
-import { Cuenta } from "./Cuenta.js";
+import { CuentaCorriente } from "./CuentaCorriente.js";
+import { CuentaAhorro } from './CuentaAhorro.js';
 
 const cliente = new Cliente("Leonardo", "123456", "123321");
 const cliente2 = new Cliente("Maria", "123142", "89899");
 
-const cuentaDeLeonardo = new Cuenta('Corriente', cliente, "1", "001", 0);
-const cuentaDeMaria = new Cuenta('Corriente', cliente2, "2", "001", 0);
+const cuentaDeLeonardo = new CuentaCorriente(cliente, "1", "001");
+const cuentaDeMaria = new CuentaCorriente(cliente2, "2", "001");
 
-const cuentaAhorroLeonardo = new Cuenta('Ahorro', cliente, '9985', '001', 0)
+const cuentaAhorroLeonardo = new CuentaAhorro(cliente, '9985', '001', 0)
 
 console.log(cuentaDeLeonardo)
 cuentaDeLeonardo.depositoEnCuenta(150)
